renaming modalOptions to optionsModal
