+++ app/components/options_modal/index.js
@@ -1,6 +1,6 @@
 // Copyright (c) 2017 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import ModalOptionsContainer from './modal_options_container';
+import OptionsModalContainer from './options_modal_container';
 
-export default ModalOptionsContainer;
+export default OptionsModalContainer;
