+++ app/reducers/views/index.js
@@ -6,13 +6,13 @@ import {combineReducers} from 'redux';
 import channel from './channel';
 import i18n from './i18n';
 import login from './login';
-import modalOptions from './modal_options';
+import optionsModal from './options_modal';
 import selectServer from './select_server';
 
 export default combineReducers({
     channel,
     i18n,
     login,
-    modalOptions,
+    optionsModal,
     selectServer
 });
